pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "system-identification"
version = "0.1.0"
description = "Flapping-wing system identification tooling"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pandas",
  "pyarrow",
  "pyulog",
  "PyYAML",
  "scipy",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
